o1vm: move memory_size into a top-level utils module
